pyproject.toml
[tool.poetry]
name = "reading-nook"
version = "0.1.0"
description = "A full service self-hostable read-it-later service"
authors = ["James Montgomery <james.montgomery@gmail.com>"]
license = "ISC"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-login = "^0.6.3"
flask-migrate = "^4.0.7"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
flask-restx = "^1.3.0"
pytest-cov = "^5.0.0"
tox = "^4.17.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
